+++ assets/js/theme-admin.js
@@ -1,0 +1,3 @@
+jQuery(function($){
+    $('.fbm-color').wpColorPicker();
+});